apu/dsp: Fold the JIT backend into the core
With one engine the DSPOps vtable, the opaque backend pointer and the
dsp_jit.c/dsp.c split only served an engine switch that no longer
exists. The core's memory, its peripheral I/O, the DMA engine's memory
callbacks and the snapshot sync now live in dsp.c, and DSPState holds
the dsp56300 handle and the three RAM buffers it maps directly.
