+++ hw/xbox/mcpx/apu/dsp/meson.build
@@ -2,7 +2,7 @@ dsp56300_emu_ffi = dependency('dsp56300-emu-ffi', fallback: ['dsp56300', 'dsp563
 
 libdsp = static_library(
   'dsp',
-  files(['dsp.c', 'dsp_jit.c', 'dsp_dma.c']) + genh,
+  files(['dsp.c', 'dsp_dma.c']) + genh,
   dependencies: [dsp56300_emu_ffi],
 )
